requirements.txt
pyzbar~=0.1.9
pillow~=9.0.1
zxing~=1.0
